+++ src/components/ChartLabel.tsx
@@ -13,23 +13,75 @@ export interface ChartLabelProps
         Animated.AnimatedProps<ViewProps> {
     alignX?: Alignment2D['x'];
     alignY?: Alignment2D['y'];
+    numberOfLines?: number;
+    // onTextLayout?: (size: { width: number; height: number }) => void;
 }
 
 const ChartLabel = (props: ChartLabelProps) => {
     const {
         alignX = 'center',
         alignY = 'center',
+        numberOfLines,
         title,
         textStyle,
         style,
         render,
+        // onTextLayout: onTextLayoutProp,
         ...otherProps
     } = props;
 
+    // const titleSize = React.useRef(new Animated.ValueXY()).current;
+    // const [titleSize, setTitleSize] = React.useState({ text: '', x: 0, y: 0 });
+
+    // const titleSizeTextRef = React.useRef('');
+
     const textProps: Animated.AnimatedProps<TextProps> = {
         selectable: false,
-        style: [{ textAlign: alignX }, textStyle],
+        style: [
+            {
+                textAlign: alignX,
+                // width: titleSize.x,
+                // height: titleSize.y,
+            },
+            // titleSize.text === title
+            //     ? {
+            //           width: titleSize.x,
+            //           height: titleSize.y,
+            //       }
+            //     : undefined,
+            textStyle,
+        ],
+        numberOfLines,
     };
+
+    // const onTextLayout = React.useCallback(
+    //     (event: NativeSyntheticEvent<TextLayoutEventData>) => {
+    //         let text = event.nativeEvent.lines.map(x => x.text).join('\n');
+    //         if (text === titleSizeTextRef.current) {
+    //             // Measure title text only once
+    //             return;
+    //         }
+    //         let width = 0;
+    //         let height = 0;
+    //         for (let line of event.nativeEvent.lines) {
+    //             if (line.width > width) {
+    //                 width = line.width;
+    //             }
+    //             height += line.height;
+    //         }
+    //         titleSizeTextRef.current = text;
+    //         onTextLayoutProp?.({
+    //             width: Math.ceil(width) + 6,
+    //             height: Math.ceil(height),
+    //         });
+    //         // setTitleSize({
+    //         //     text,
+    //         //     x: Math.ceil(width) + 6,
+    //         //     y: Math.ceil(height),
+    //         // });
+    //     },
+    //     [onTextLayoutProp]
+    // );
 
     let content: React.ReactNode;
     if (render) {
@@ -40,7 +92,14 @@ const ChartLabel = (props: ChartLabelProps) => {
             );
         }
     } else {
-        content = <Animated.Text {...textProps}>{title || ''}</Animated.Text>;
+        content = (
+            <Animated.Text
+                {...textProps}
+                // onTextLayout={Platform.OS !== 'web' ? onTextLayout : undefined}
+            >
+                {title || ''}
+            </Animated.Text>
+        );
     }
 
     return (